Added constructors to PlayerCrsh for setting up lives and coordinates as wanted, added some inheritance between CircleComponent and DrawableComponent and overall comment fixes

+++ app/src/main/java/com/example/tomascrd/c_r_s_h/components/DrawableComponent.java
@@ -17,8 +17,8 @@ public abstract class DrawableComponent {
     /**
      * X and Y coordinates
      */
-    protected int xPos;
-    protected int yPos;
+    protected float xPos;
+    protected float yPos;
 
     /**
      * Draws the components on screen
